03- ML Exp with DVC pipeline/requirements.txt
scikit-learn==1.2.2
sklearn-features==0.0.2
imblearn
matplotlib
seaborn
joblib
dvc